fix: removing payment-option-list component and inserting this list at payment-method page
