+++ src/pages/PaymentMethod/index.tsx
@@ -1,4 +1,4 @@
-import PaymentOptionList from "../../components/PaymentOptionList";
+import { useState } from "react";
 import PaymentOptionCard from "../../components/ui/PaymentOptionCard";
 import Title from "../../components/ui/Title";
 import TopLabel from "../../components/ui/TopLabel";
@@ -7,6 +7,12 @@ import "./index.css";
 
 const PaymentMethod = () => {
   const firstInstallment = data.paymentMethods.installments[0];
+  const installments = data.paymentMethods.installments.slice(1);
+  const [checkedItem, setCheckedItem] = useState<number | null>(null);
+
+  const handleCheckboxChange = (id: number | null, checked: boolean) => {
+    setCheckedItem(checked ? id : null);
+  };
 
   return (
     <>
@@ -25,10 +31,38 @@ const PaymentMethod = () => {
             benefitSuffix={firstInstallment.benefitSuffix}
             highlight={firstInstallment.highlight}
             highlightValue={firstInstallment.highlightValue}
+            checked={checkedItem === 0}
+            onChange={(checked) => handleCheckboxChange(0, checked)}
           />
         </div>
       </div>
-      <PaymentOptionList />
+      <div className="payment-option-list">
+        <TopLabel label="Pix Parcelado" />
+        {installments.map((installment) => (
+          <div
+            key={installment.id}
+            className={`payment-option-item ${
+              checkedItem === installment.id ? "checked" : ""
+            }`}
+          >
+            <PaymentOptionCard
+              installments={installment.label}
+              amount={installment.amount}
+              paymentInstallment={installment.paymentInstallment}
+              benefitPrefix={installment.benefitPrefix || ""}
+              benefitPercentage={installment.benefitPercentage || ""}
+              benefitSuffix={installment.benefitSuffix || ""}
+              highlight={installment.highlight || ""}
+              highlightValue={installment.highlightValue || ""}
+              useListCard={true}
+              checked={checkedItem === installment.id}
+              onChange={(checked) =>
+                handleCheckboxChange(installment.id, checked)
+              }
+            />
+          </div>
+        ))}
+      </div>
     </>
   );
 };
